Anubis yum repos management configuration
First, all files have to belong to root:root and have 644 permissions.

Second, if puppet applies a configuration change, it has to re-generate the
metadata. In the current scenario, when files are moved into the repo with the
user / group that uploaded them, it makes sense. A better way possibly needs to
be found.