Make Postfix listen on an internal virtual interface
Unfortunately, virbr0 is brought up by libvirt later than Postfix
starts, so it can't bind to this interface.

However, it is undesirable to have it listen an external interface as
well, for which reason a virtual interface is introduced that would not
be accesible to the outside world and yet accessible to the virtual
machines.